+++ include/arviss/rv32_concepts.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "arviss/core_concepts.h"
+#include "arviss/core/concepts.h"
 #include "arviss/types.h"
 
 #include <concepts>
