+++ include/arviss/rv32_dispatchers.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "arviss/core_concepts.h"
+#include "arviss/core/concepts.h"
 #include "arviss/instruction.h"
 #include "arviss/rv32_concepts.h"
