+++ include/arviss/rv32_executors.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "arviss/core_concepts.h"
+#include "arviss/core/concepts.h"
 #include "arviss/rv32_concepts.h"
 
 #include <bit>
